backend/requirements.txt
django==5.0.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
neomodel==5.3.*
django-cors-headers==4.3.*
cloudinary==1.40.*
argon2-cffi==23.1.*
bleach==6.1.*
python-slugify==8.0.*
python-decouple==3.8
Pillow==10.*
gunicorn==22.*
whitenoise==6.7.*
psycopg2-binary==2.9.*
